Cleaned up some compile issues on unix

+++ src/ZoneServer/AbstractState.h
@@ -36,7 +36,7 @@ enum StateTypes
     State_Locomotion = 2
 };
 typedef std::vector<uint64> stateRemove;
-typedef std::map<StateTypes, uint64> transitionList;
+typedef std::map<StateTypes, uint32> transitionList;
 class IState
 {
 public:
